pyproject.toml
[project]
name = "chatbot-condominium"
version = "0.1.0"
description = ""
authors = [
    {name = "Murilo Cedraz"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "python-telegram-bot (>=21.11.1,<22.0.0)",
    "pymupdf (>=1.25.3,<2.0.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "langchain-openai (>=0.3.8,<0.4.0)",
    "openai (>=1.66.3,<2.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "langchain-community (>=0.3.19,<0.4.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "langchain-google-genai (>=2.1.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
